requirements.txt
streamlit
openai
PyYAML
networkx
